# backend/openapi.yml
openapi: 3.0.3
info:
  title: YouSightseeing API
  description: API для туристического приложения (поиск мест, построение маршрутов, профиль)
  version: "1.0.0"

tags:
  - name: Auth
    description: Авторизация и профиль
  - name: Places
    description: Поиск достопримечательностей (прокси к Geopify)
  - name: Routes
    description: Маршруты пользователя

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # --- ОШИБКИ ---
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid token"
    
    # --- МОДЕЛИ ---
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        google_sub:
          type: string
          example: "10769150350006150715113082367"
        email:
          type: string
          example: "kirpich@example.com"
        full_name:
          type: string
          example: "Kirill Kirpichov"
        picture:
          type: string
          format: uri
          example: "https://lh3.googleusercontent.com/a-/AOh14Hg..."
        first_name:
          type: string
          example: "Kirill"
        last_name:
          type: string
          example: "Kirpichov"
        email_verified:
          type: boolean
          example: true
        google_domain:
          type: string
          example: "example.com"
        locale:
          type: string
          example: "ru"
        created_at:
          type: string
          format: date-time
          example: "2025-12-08T18:44:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-12-09T10:00:00Z"

    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: "Действующий refresh токен пользователя"
          example: "d9b2d63d-a233-4323-83d0-3b6c2d1c5a1d"

    RefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: "JWT токен доступа"
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          description: "Новый refresh токен"
          example: "8a6e0804-24d6-4470-bf21-23588921d7b3"
        user:
          $ref: '#/components/schemas/User'

    Place:
      type: object
      required: [name, lat, lon, category]
      properties:
        name:
          type: string
          example: "Tretyakov Gallery"
        lat:
          type: number
          format: double
          example: 55.741
        lon:
          type: number
          format: double
          example: 37.620
        category:
          type: string
          enum: [museum, park, cafe, monument, other]
        address:
          type: string
          nullable: true
        
    Route:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "Прогулка по центру"
        points:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        duration_seconds:
          type: integer
          example: 3600
        distance_meters:
          type: integer
          example: 5000
        created_at:
          type: string
          format: date-time

paths:
  # ==========================
  # АВТОРИЗАЦИЯ
  # ==========================
  /auth/google:
    post:
      tags: [Auth]
      summary: Вход через Google
      description: "Создает нового пользователя или авторизует существующего через Google ID Token"
      operationId: authGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [google_token]
              properties:
                google_token:
                  type: string
                  description: "Токен, полученный от Google Sign-In на клиенте"
                  example: "eyJhbGciOiJSUzI1NiIsImtpZCI6Im..."
      responses:
        '201':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          description: Неверный формат запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Токен Google невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление токенов
      description: "Получает новый Access Token по валидному Refresh Token"
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Токены успешно обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Refresh токен невалиден, истек или отозван
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход из системы
      description: "Отзывает указанный Refresh Token, делая его невалидным."
      operationId: authLogout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  description: "Refresh токен, который нужно отозвать"
                  example: "8a6e0804-24d6-4470-bf21-23588921d7b3"
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  logout:
                    type: boolean
                    example: true
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /user/profile:
    get:
      tags: [Auth]
      summary: Получить профиль текущего пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==========================
  # ПОИСК МЕСТ (PROXY)
  # ==========================
  /places/search:
    get:
      tags: [Places]
      summary: Найти места рядом (через Geopify)
      description: Бэкэнд сам сходит в Geopify API и вернет чистый список
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lon
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: radius
          in: query
          description: Радиус поиска в метрах (по умолчанию 1000)
          schema:
            type: integer
            default: 1000
        - name: categories
          in: query
          description: Категории через запятую (museum,park,cafe)
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  places:
                    type: array
                    items:
                      $ref: '#/components/schemas/Place'

  # ==========================
  # МАРШРУТЫ
  # ==========================
  /routes:
    get:
      tags: [Routes]
      summary: Получить список сохраненных маршрутов
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Route'

    post:
      tags: [Routes]
      summary: Сохранить маршрут
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, points]
              properties:
                name:
                  type: string
                points:
                  type: array
                  items:
                    $ref: '#/components/schemas/Place'
                duration:
                  type: integer
                distance:
                  type: integer
      responses:
        '201':
          description: Маршрут сохранен
          content:
            application/json:
              schema:
                type: object
                properties:
                  route_id:
                    type: integer

  /routes/{id}:
    delete:
      tags: [Routes]
      summary: Удалить маршрут
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешно удалено
        '404':
          description: Маршрут не найден